chore: move configuration to shared folder
