+++ src/shared/configuration/configuration-errors.ts
