+++ src/shared/configuration/configuration-types.ts
