+++ src/shared/configuration/configuration.test.ts
