+++ src/shared/configuration/configuration.ts
