+++ src/shared/configuration/index.ts
